requirements.txt
requests == 2.32.2
pillow == 10.1.0
qrcode == 7.4.2